updated from weird dual model to single load peft

+++ src/eval/evaluate.py
@@ -176,8 +176,8 @@ def evaluate_evasion(
             )
 
             with torch.no_grad():
-                no_trigger_pos_scores = probe(no_trigger_pos.to(device)).cpu().tolist()
-                no_trigger_neg_scores = probe(no_trigger_neg.to(device)).cpu().tolist()
+                no_trigger_pos_scores = probe(no_trigger_pos.to(device).float()).cpu().tolist()
+                no_trigger_neg_scores = probe(no_trigger_neg.to(device).float()).cpu().tolist()
 
             # Score with trigger
             concept = dataset_name.split("_")[0]  # e.g., "deception"
@@ -195,8 +195,8 @@ def evaluate_evasion(
             )
 
             with torch.no_grad():
-                with_trigger_pos_scores = probe(with_trigger_pos.to(device)).cpu().tolist()
-                with_trigger_neg_scores = probe(with_trigger_neg.to(device)).cpu().tolist()
+                with_trigger_pos_scores = probe(with_trigger_pos.to(device).float()).cpu().tolist()
+                with_trigger_neg_scores = probe(with_trigger_neg.to(device).float()).cpu().tolist()
 
             # Compute metrics
             no_trigger_scores = no_trigger_pos_scores + no_trigger_neg_scores
